+++ timezonefinder/helpers_numba.py
@@ -1,6 +1,9 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 from math import asin, atan2, ceil, cos, degrees, radians, sin, sqrt
+
+from six.moves import range
+
 from numba import b1, f8, i2, i4, jit, typeof, u2, u8
 
 # # for Ahead-Of-Time Compilation:
@@ -8,6 +11,8 @@ from numba import b1, f8, i2, i4, jit, typeof, u2, u8
 # cc = CC('compiled_helpers', )
 # # Uncomment the following line to print out the compilation steps
 # # cc.verbose = True
+
+TIMEZONE_NAMES_FILE = 'timezone_names.json'
 
 dtype_3floattuple = typeof((1.0, 1.0, 1.0))
 dtype_2floattuple = typeof((1.0, 1.0))