+++ CHANGELOG.rst
@@ -7,6 +7,22 @@ Changelog
 * bugfix: possibly uninitialized variable in closest_timezone_at()
 
 
+3.0.0 (2018-05-17)
+------------------
+
+* ATTENTION: the package six is now required! (was necessary because of the new testing routine. improves compatibility standards)
+* updated build/testing/publishing routine
+* updated the data to `2018d <https://github.com/evansiroky/timezone-boundary-builder/releases/tag/2018d>`__
+* fixing minor issue #52 (shortcuts being out of bounds for extreme coordinate values)
+* the list of polygon ids in each shortcut is sorted after freq. of appearance of their zone id.
+    this is critical for ruling out zones faster (as soon as just polygons of one zone are left this zone can be returned)
+* using argparse package now for parsing the command line arguments
+* added option of choosing between functions timezone_at() and certain_timezone_at() on the command line with flag -f
+* the timezone names are now being stored in a readable JSON file
+* adjusted the main test cases
+* corrections and clarifications in the readme and code comments
+
+
 2.1.1 (2017-11-20)
 ------------------